+++ agent/agent-instrumentation/src/main/java/org/bithon/agent/instrumentation/aop/interceptor/InterceptorManager.java
@@ -18,15 +18,13 @@ package org.bithon.agent.instrumentation.aop.interceptor;
 
 
 import org.bithon.agent.instrumentation.aop.interceptor.declaration.AbstractInterceptor;
-import org.bithon.agent.instrumentation.loader.PluginClassLoaderManager;
-import org.bithon.agent.instrumentation.logging.ILogger;
 import org.bithon.agent.instrumentation.logging.LoggerFactory;
 
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.function.Supplier;
 
 /**
  * Interceptor is singleton
@@ -35,11 +33,9 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public class InterceptorManager {
 
-    private static final ILogger LOG = LoggerFactory.getLogger(InterceptorManager.class);
-
-    private int interceptorIndex = 0;
-
     public static final InterceptorManager INSTANCE = new InterceptorManager();
+
+    private int globalInterceptorIndex = 0;
 
     private InterceptorSupplier[] interceptorList = new InterceptorSupplier[64];
 
@@ -54,69 +50,14 @@ public class InterceptorManager {
     /**
      * Get interceptor by given index.
      */
-    public InterceptorSupplier getSupplier(int index) {
-        return index < interceptorList.length ? interceptorList[index] : null;
+    public Supplier<AbstractInterceptor> getSupplier(int index) {
+        return index < interceptorList.length ? interceptorList[index] : () -> null;
     }
 
-    public static class InterceptorSupplier {
-        private String interceptorClassName;
-        private ClassLoader classLoader;
-
-        private volatile AbstractInterceptor interceptor;
-        private final int index;
-        private boolean initialized = false;
-
-        public InterceptorSupplier(int index,
-                                   String interceptorClassName,
-                                   ClassLoader classLoader) {
-            this.index = index;
-            this.interceptorClassName = interceptorClassName;
-            this.classLoader = classLoader;
-        }
-
-        public boolean isInitialized() {
-            return initialized;
-        }
-
-        public AbstractInterceptor get() {
-            if (interceptor != null) {
-                return interceptor;
-            }
-
-            synchronized (this) {
-                // Double check
-                if (interceptor != null) {
-                    return interceptor;
-                }
-
-                interceptor = createInterceptor(interceptorClassName, classLoader);
-                if (interceptor != null) {
-                    // Release reference so that GC works
-                    this.classLoader = null;
-                    this.interceptorClassName = null;
-                }
-                initialized = true;
-                return interceptor;
-            }
-        }
-
-        private AbstractInterceptor createInterceptor(String interceptorClassName, ClassLoader userClassLoader) {
-            try {
-                // Load class out of lock in case of deadlock
-                ClassLoader interceptorClassLoader = PluginClassLoaderManager.getClassLoader(userClassLoader);
-                Class<?> interceptorClass = Class.forName(interceptorClassName, true, interceptorClassLoader);
-
-                return (AbstractInterceptor) interceptorClass.getConstructor().newInstance();
-            } catch (Throwable e) {
-                LOG.error(String.format(Locale.ENGLISH,
-                                        "Failed to load interceptor[%s] due to %s",
-                                        interceptorClassName,
-                                        e.getMessage()), e);
-                return null;
-            }
-        }
-    }
-
+    /**
+     * Get or create an interceptor supplier
+     * @return the global index of this supplier so that it can be passed to {{@link #getSupplier(int)}} to get the supplier instance
+     */
     public int getOrCreateSupplier(String interceptorClassName,
                                    ClassLoader classLoader) {
         Map<String, InterceptorSupplier> suppliers = interceptorMaps.computeIfAbsent(interceptorClassName, (v) -> new ConcurrentHashMap<>());
@@ -124,18 +65,18 @@ public class InterceptorManager {
         String classLoaderId = classLoader == null ? "bootstrap" : classLoader.getClass().getName() + "@" + System.identityHashCode(classLoader);
         InterceptorSupplier supplier = suppliers.get(classLoaderId);
         if (supplier != null) {
-            return supplier.index;
+            return supplier.getIndex();
         }
 
         synchronized (InterceptorManager.class) {
             // Double check
             supplier = suppliers.get(classLoaderId);
             if (supplier != null) {
-                return supplier.index;
+                return supplier.getIndex();
             }
 
-            ensureCapacity(interceptorIndex);
-            int index = interceptorIndex++;
+            ensureCapacity(globalInterceptorIndex);
+            int index = globalInterceptorIndex++;
 
             supplier = new InterceptorSupplier(index, interceptorClassName, classLoader);
             suppliers.put(classLoaderId, supplier);
@@ -143,7 +84,7 @@ public class InterceptorManager {
             interceptorList[index] = supplier;
         }
 
-        return supplier.index;
+        return supplier.getIndex();
     }
 
     /**
@@ -161,6 +102,6 @@ public class InterceptorManager {
         InterceptorSupplier[] newArray = new InterceptorSupplier[(int) (interceptorList.length * 1.5)];
         System.arraycopy(interceptorList, 0, newArray, 0, interceptorList.length);
         interceptorList = newArray;
-        LOG.info("Enlarge dynamic interceptors storage to {}", interceptorList.length);
+        LoggerFactory.getLogger(InterceptorManager.class).info("Enlarge dynamic interceptors storage to {}", interceptorList.length);
     }
 }
